+++ desktop/qml/PhotoPane.qml
@@ -1,24 +1,34 @@
+/*
+ * Copyright (C) 2018-2019 André Hoarau <ah@ouvaton.org>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; version 3.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 import QtQuick 2.10
 import QtQuick.Controls 2.4
-import QtQuick.Layouts 1.3
 import QtQuick.Controls.Material 2.3
-import Qt.labs.settings 1.0
-import Qt.labs.platform 1.0 as Platform
-import QtQuick.Window 2.10
-
-import io.qrop.components 1.0
 
 Pane {
     id: photoPane
     
     property var photoIdList
+    padding: 0
     
     onPhotoIdListChanged: {
         photoModel.clear();
         for (var i = 0; i< photoIdList.length; i++) {
             photoModel.append({"photoId": photoIdList[i]})
         }
-        
     }
     
     ListModel {
@@ -26,20 +36,29 @@ Pane {
     }
 
     ToolButton {
+        id: closeButton
         text: "\ue14c"
         font.family: "Material Icons"
         font.pixelSize: Units.fontSizeHeadline
         onClicked: photoPane.visible = false
         anchors.top: parent.top
         anchors.right: parent.right
-        anchors.margins: -padding
+        anchors.margins: 16 -padding
+        Material.foreground: Units.closeButtonColor
     }
 
     SwipeView {
         id: swipeView
-        anchors.centerIn: parent
-        width: parent.width * 0.9
-        height: parent.height * 0.9
+        anchors {
+            top: closeButton.bottom
+            left: parent.left
+            right: parent.right
+            bottom: indicator.top
+            margins: Units.largeSpacing
+        }
+
+        clip: true
+        currentIndex: indicator.currentIndex
 //        transitions: null
         Repeater {
             model: photoModel
@@ -84,8 +103,12 @@ Pane {
         
         count: swipeView.count
         currentIndex: swipeView.currentIndex
+        interactive: true
         
-        anchors.top: swipeView.bottom
-        anchors.horizontalCenter: parent.horizontalCenter
+        anchors {
+            bottom: parent.bottom
+            bottomMargin: Units.formSpacing
+            horizontalCenter: parent.horizontalCenter
+        }
     }
 }